build(win64): port the crash reporters off x86, and five smaller x64 blockers
With the include-path and stub problems out of the way the x64 build reaches
1752 of 1778 objects, and what is left is six causes across sixteen steps.

The big one is the crash-reporting code - WWLib/Except.cpp, both copies of
GameEngine/Source/Common/System/StackDump.cpp, and Libraries/Source/debug's
debug_except / debug_stack / debug_debug. All of it was written for 32-bit x86
and hardcodes three things that do not exist at x64:

  * CONTEXT::Eip / Esp / Ebp. The x64 record calls them Rip / Rsp / Rbp.
  * MSVC inline assembly to read those registers. The x64 compiler rejects
    __asm outright (C4235), and the leftover tokens then produce a cascade of
    "'eax': undeclared identifier".
  * IMAGE_FILE_MACHINE_I386, plus DWORD-wide dbghelp signatures. dbghelp.h
    quietly remaps StackWalk, STACKFRAME, IMAGEHLP_SYMBOL, SymGetModuleBase and
    friends to their ...64 forms under _WIN64, and those take DWORD64
    addresses - hence the "cannot convert argument 7" errors.

Dependencies/Utility/Utility/cpu_context_compat.h now supplies GX_CTX_PC/SP/FP,
the native StackWalk machine type, and GXCaptureContextRegisters(), which uses
RtlCaptureContext instead of assembly. That works on x86 too, so all three
former code paths collapse into one. Address parameters through DbgHelpLoader,
Except.cpp and debug_stack.inl become DWORD_PTR, which is DWORD at 32-bit, so
the win32 build is unchanged.

Two latent 64-bit bugs turned up on the way and are fixed here:
dbghelp entry points were looked up by their unsuffixed names, which a 64-bit
dbghelp.dll does not export at all, and debug_stack stored those pointers in an
`unsigned`, discarding the top half of every address. Both would have left the
stack-trace facility silently dead rather than merely inaccurate.

Known limitation, called out in the header: Signature keeps its 32-bit address
array, so x64 stack signatures still truncate. That is a symbol-resolution
quality problem in a diagnostic path, not a correctness one, and widening it
reaches into the profiler.

The five smaller causes:

  * Generals' BezierSegment.h tested _WIN32 before SAGE_USE_GLM. On Windows
    both are defined, so it included d3dx8math.h, skipped glm, and then
    declared a glm::mat4 member. Zero Hour's copy already had the arms the
    other way round.
  * D3DXBUFFER was a bare forward declaration, but W3DWater.cpp calls
    GetBufferPointer() and Release() on one. Given a body matching
    ID3DXBuffer; nothing constructs one, since D3DXAssembleShader is a stub.
  * IMEManager.cpp included mbstring_compat.h, which lives in the POSIX shim
    directory that is deliberately off the Windows include path now. Its
    Windows arm is "#include <mbstring.h>", so do that directly.
  * TheSystemIsUnicode was read by Core's Debug.cpp but declared nowhere, and
    defined only by Zero Hour. Declared where it is used, defined for Generals
    too.
  * GeneralsMD's GameEngine.cpp walks the save directory with opendir/readdir.
    Added Utility/dirent_compat.h, a Windows-only shim over _findfirst, rather
    than forking the walk.

Both new headers are inside #ifdef _WIN32 in their entirety. Every source
change is either inside an existing _WIN32 guard or resolves to the previous
behaviour off Windows. macOS and iOS both build clean, exit 0, no "error:".

+++ GeneralsMD/Code/GameEngine/Include/Common/StackDump.h
@@ -35,7 +35,10 @@ void StackDump(void (*callback)(const char*));
 
 // Writes a stackdump (provide a callback : gets called per line)
 // If callback is nullptr then will write using OuputDebugString
-void StackDumpFromContext(DWORD eip,DWORD esp,DWORD ebp, void (*callback)(const char*));
+// GeneralsX @bugfix Claude 27/07/2026 Pointer-width, not DWORD: at x64 an instruction
+// pointer does not fit in 32 bits. size_t rather than DWORD_PTR because this header is
+// also parsed by the macOS, iOS and Linux builds, where DWORD_PTR does not exist.
+void StackDumpFromContext(size_t eip,size_t esp,size_t ebp, void (*callback)(const char*));
 
 // Gets count* addresses from the current stack
 void FillStackAddresses(void**addresses, unsigned int count, unsigned int skip = 0);
